moved 3 pages to components. changed order
